refactor: replace wallet store usage with global authentication store across components
