+++ src/components/modules/escrow/ui/cards/MyEscrowsCards.tsx
@@ -8,7 +8,7 @@ import { Input } from "@/components/ui/input";
 import NoData from "@/components/utils/NoData";
 import EscrowDetailDialog from "../dialogs/EscrowDetailDialog";
 import { useEscrowBoundedStore } from "../../store/ui";
-import { useGlobalBoundedStore } from "@/core/store";
+import { useGlobalBoundedStore } from "@/core/store/data";
 import LoaderData from "@/components/utils/LoaderData";
 
 interface MyEscrowsCardsProps {
